requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.9.24
charset-normalizer==2.1.1
Deprecated==1.2.13
distro==1.8.0
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.2
netaddr==0.8.0
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
urllib3==1.26.13
wrapt==1.14.1
yarl==1.8.1
yookassa==2.3.2
